+++ services/supervisor/tests/supervisor/main.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use exonum_merkledb::ObjectHash;
-use exonum_testkit::{ApiKind, InstanceCollection, TestKit, TestKitApi, TestKitBuilder};
+use exonum_testkit::{ApiKind, TestKit, TestKitApi, TestKitBuilder};
 
 use exonum::{
     api,
@@ -217,20 +217,24 @@ fn start_service_instance(testkit: &mut TestKit, instance_name: &str) -> Instanc
 fn testkit_with_inc_service() -> TestKit {
     TestKitBuilder::validator()
         .with_logger()
-        .with_rust_service(Supervisor::builtin_instance(
+        .with_rust_service(Supervisor)
+        .with_artifact(Supervisor.artifact_id())
+        .with_instance(Supervisor::builtin_instance(
             Supervisor::decentralized_config(),
         ))
-        .with_rust_service(InstanceCollection::new(IncService))
+        .with_rust_service(IncService)
         .create()
 }
 
 fn testkit_with_inc_service_and_n_validators(n: u16) -> TestKit {
     TestKitBuilder::validator()
         .with_logger()
-        .with_rust_service(Supervisor::builtin_instance(
+        .with_rust_service(Supervisor)
+        .with_artifact(Supervisor.artifact_id())
+        .with_instance(Supervisor::builtin_instance(
             Supervisor::decentralized_config(),
         ))
-        .with_rust_service(InstanceCollection::new(IncService))
+        .with_rust_service(IncService)
         .with_validators(n)
         .create()
 }
@@ -242,30 +246,30 @@ fn testkit_with_inc_service_and_two_validators() -> TestKit {
 fn testkit_with_inc_service_auditor_validator() -> TestKit {
     TestKitBuilder::auditor()
         .with_logger()
-        .with_rust_service(Supervisor::builtin_instance(
+        .with_rust_service(Supervisor)
+        .with_artifact(Supervisor.artifact_id())
+        .with_instance(Supervisor::builtin_instance(
             Supervisor::decentralized_config(),
         ))
-        .with_rust_service(InstanceCollection::new(IncService))
+        .with_rust_service(IncService)
         .with_validators(1)
         .create()
 }
 
 fn testkit_with_inc_service_and_static_instance() -> TestKit {
-    let service = IncService;
-    let collection = InstanceCollection::new(service).with_instance(SERVICE_ID, SERVICE_NAME, ());
     TestKitBuilder::validator()
         .with_logger()
-        .with_rust_service(Supervisor::builtin_instance(
+        .with_rust_service(Supervisor)
+        .with_artifact(Supervisor.artifact_id())
+        .with_instance(Supervisor::builtin_instance(
             Supervisor::decentralized_config(),
         ))
-        .with_rust_service(collection)
+        .with_default_rust_service(IncService)
         .create()
 }
 
 fn add_available_services(runtime: RustRuntime) -> RustRuntime {
-    runtime
-        .with_available_service(IncService)
-        .with_available_service(Supervisor)
+    runtime.with_factory(IncService).with_factory(Supervisor)
 }
 
 /// Just test that the Inc service works as intended.
@@ -564,10 +568,12 @@ fn test_start_two_services_in_one_request() {
 fn test_restart_node_and_start_service_instance() {
     let mut testkit = TestKitBuilder::validator()
         .with_logger()
-        .with_rust_service(Supervisor::builtin_instance(
+        .with_rust_service(Supervisor)
+        .with_artifact(Supervisor.artifact_id())
+        .with_instance(Supervisor::builtin_instance(
             Supervisor::decentralized_config(),
         ))
-        .with_rust_service(InstanceCollection::new(IncService))
+        .with_rust_service(IncService)
         .create();
     deploy_default(&mut testkit);
 
@@ -1031,18 +1037,20 @@ fn test_id_assignment() {
 #[test]
 fn test_id_assignment_sparse() {
     let max_builtin_id = 100;
+    let inc_service = IncService;
+    let inc_service_artifact = inc_service.artifact_id();
 
     // Create testkit with builtin instance with ID 100.
     let mut testkit = TestKitBuilder::validator()
         .with_logger()
-        .with_rust_service(Supervisor::builtin_instance(
+        .with_rust_service(Supervisor)
+        .with_artifact(Supervisor.artifact_id())
+        .with_instance(Supervisor::builtin_instance(
             Supervisor::decentralized_config(),
         ))
-        .with_rust_service(InstanceCollection::new(IncService).with_instance(
-            max_builtin_id,
-            "inc",
-            (),
-        ))
+        .with_artifact(inc_service_artifact.clone())
+        .with_instance(inc_service_artifact.into_default_instance(max_builtin_id, "inc"))
+        .with_rust_service(inc_service)
         .create();
 
     let artifact = artifact_default();